Update article on PDF Table Of Contents bookmarks re LibreOffice
LibreOffice does not work unless the exported file GoogleDocs is both
exported as a .docx and another program like Word has fixed things in
the file (somehow).  Since this means you have to use a tool that
by itself solves the problem just to use LibreOffice as part of a longer
solution, it is not worth doing.